Rename shared module to auth
