+++ androidNekomp/build.gradle.kts
@@ -5,10 +5,10 @@ plugins {
 }
 
 android {
-    namespace = "com.chesire.nekomp.android"
+    namespace = "com.chesire.nekomp"
     compileSdk = 35
     defaultConfig {
-        applicationId = "com.chesire.nekomp.android"
+        applicationId = "com.chesire.nekomp"
         minSdk = 27
         targetSdk = 35
         versionCode = 1
@@ -27,17 +27,10 @@ android {
             isMinifyEnabled = false
         }
     }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
 }
 
 dependencies {
-    implementation(projects.shared)
+    implementation(projects.library.datasource.auth)
     implementation(libs.compose.ui)
     implementation(libs.compose.ui.tooling.preview)
     implementation(libs.compose.material3)
