+++ androidNekomp/src/main/java/com/chesire/nekomp/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekomp.android
+package com.chesire.nekomp
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.chesire.nekomp.Greeting
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
